[REFACTOR]<Attempt> : BigInteger 변수를 int로 변경하여 성능 및 메모리 최적화 - BigInteger를 int로 변경하여 불필요한 메모리 사용을 줄이고 성능을 개선 - 값의 범위가 int로 충분하다고 판단하여 자료형을 단순화함

+++ src/main/java/racingcar/domain/Car.java
@@ -17,8 +17,8 @@ public class Car {
         return name.getName();
     }
 
-    public BigInteger getPosition() {
-        return BigInteger.valueOf(position.positionDistance());
+    public int getPosition() {
+        return position.positionDistance();
     }
 
     public String currentPositionFormat() {
